6. RegisterCommentDTO 분리
