+++ src/main/java/com/together/buytogether/postcomment/feature/RegisterComment.java
@@ -7,13 +7,10 @@ import com.together.buytogether.post.domain.Post;
 import com.together.buytogether.post.domain.PostRepository;
 import com.together.buytogether.postcomment.domain.PostComment;
 import com.together.buytogether.postcomment.domain.PostCommentRepository;
+import com.together.buytogether.postcomment.dto.request.RegisterCommentDTO;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-
-import java.time.LocalDateTime;
 
 @RestController
 public class RegisterComment {
@@ -35,30 +32,11 @@ public class RegisterComment {
     public void request(
             @SessionAttribute(name = SessionConst.LOGIN_MEMBER) Long memberId,
             @PathVariable Long postId,
-            @RequestBody @Valid Request request) {
+            @RequestBody @Valid RegisterCommentDTO registerCommentDTO) {
         Member member = memberRepository.getByMemberId(memberId);
         Post post = postRepository.getByPostId(postId);
-        PostComment comment = request.toDomain(member, post);
+        PostComment comment = registerCommentDTO.toDomain(member, post);
         postCommentRepository.save(comment);
-    }
-
-    public record Request(
-            @NotBlank(message = "댓글 내용은 필수입니다")
-            String content,
-            @NotNull(message = "댓글 생성일은 필수입니다")
-            LocalDateTime createAt,
-            @NotNull(message = "댓글 수정일은 필수입니다")
-            LocalDateTime updateAt
-    ) {
-        public PostComment toDomain(Member member, Post post) {
-            return new PostComment(
-                    member,
-                    post,
-                    content,
-                    createAt,
-                    updateAt
-            );
-        }
     }
 
 }
